Implement a "Compress no" option to disable IMAP compression
As noted in the documentation this is mainly for debugging. I have a
fickle IMAP server that was throwing errors on compression, and as noted
in the documentation now I can use this in conjunction with stunnel to
see what's going on with it.

When I was searching for a way to do this I found
http://sourceforge.net/p/isync/mailman/message/33285999/ which noted
that this was easy to change but there was no option for it, now we have
that option.