Document how to use DisableExtensions to disable compression
See "[PATCH] Add an option to disable compression with IMAP". I
initially implemented this as a new option, but missed that you can use
the DisableExtensions option for this.

Keep the documentation I had in the initial patch reworded so that it
applies for the DisableExtensions option, this ensures that anyone
searching for "compression" in the manpage will find this.